move tests back to main file
